refactor: Comment 및 Reply CQRS 구조 분리
- 상위 도메인 Repository 직접 참조 제거 (Comment → Post 등)
- 상위 도메인 쿼리 서비스(PostQueryService 등)에서 유효성 검사 처리
- 의존성 순환 문제 발생 → Query/Command 분리로 해결
