+++ src/main/java/inu/codin/codin/domain/post/domain/comment/repository/CommentRepository.java
@@ -6,10 +6,12 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
 
+@Repository
 public interface CommentRepository extends MongoRepository<CommentEntity, ObjectId> {
